Upgrade to latest macroquad. Try matching bundled js. Still no wasm.
Loading texture does work ok. It's rendering it which is the problem.

And latest macroquad + bundled js DOES load ok, whereas previously it
failed to load. Presumably a version mismatch with one of the other
libraries in the bundle as I had an old macroquad.

Suddenly realise, the game is still loading textures each frame, it
probably runs out of resources. That I never deallocated them because
I imagined them being cleared at end of game. Should try not doing that.